refactor: removed unused components and moved common files to common example folders and format
