+++ apps/common-app/src/components/index.ts
@@ -1,6 +1,4 @@
 export { default as Icon } from './Icon';
-export { default as Step } from './Step';
-export { default as Steps } from './Steps';
 export { default as Button } from './Button';
 export { default as Slider } from './Slider';
 export { default as Spacer } from './Spacer';
