+++ apps/common-app/src/types.ts
@@ -1,8 +1,39 @@
-export type SoundName = 'kick' | 'hi-hat' | 'clap' | 'downbeat' | 'regularbeat';
+export type InstrumentName = 'kick' | 'clap' | 'hi-hat';
 
-export type SoundRow = {
-  name: SoundName;
-  steps: boolean[];
-};
+export interface Instrument {
+  color: string;
+  radius: number;
+  name: InstrumentName;
+}
 
-export type Sounds = SoundRow[];
+export interface Touché {
+  cX: number;
+  cY: number;
+  stepIdx: number;
+  patternIdx: number;
+}
+
+export interface Pattern {
+  instrumentName: InstrumentName;
+  steps: Array<boolean>;
+}
+
+export interface Preset {
+  name: string;
+  bpm: number;
+  pattern: Pattern[];
+}
+
+export interface XYWHRect {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
+
+export interface XYPoint {
+  x: number;
+  y: number;
+}
+
+export type PlayingInstruments = Record<InstrumentName, boolean>;
